hw6.Edit3

+++ homework6/App/Controllers/AdminPanel/Delete.php
@@ -7,14 +7,17 @@ use App\AdminPanelController;
 class Delete extends AdminPanelController
 {
     /**
-     * @throws \App\Exceptions\Db
-     * @throws \App\Exceptions\DbNotFoundRecord
+     * @throws \App\Exceptions\DbException
+     * @throws \App\Exceptions\E404Exception
      */
     protected function action()
     {
         if (isset($_GET['id'])) {
             $article = \App\Models\Article::findById($_GET['id']); //Article
-            $article->delete();
+
+            if (false !== $article) {
+                $article->delete();
+            }
         }
         header('Location: /homework6/adminPanel/index/');
     }
